ServerLoadBack/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
firebase-admin
